註冊
輸入註冊資料會傳至資料庫

+++ workspace/springbootdemo/src/main/java/com/loveGod/demo/model/RegisterDao.java
@@ -2,6 +2,8 @@ package com.loveGod.demo.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegisterDao extends JpaRepository<RegisterModel, Long> {
+public interface RegisterDao extends JpaRepository<RegisterModel, Integer> {
+
+
 
 }
